(PC-42565) refactor(home): use artist query to display artist information on artist playlist module
