+++ src/features/artist/pages/Artist.tsx
@@ -4,11 +4,11 @@ import { ErrorBoundary } from 'react-error-boundary'
 import { ViewToken } from 'react-native'
 
 import { ArtistBody } from 'features/artist/components/ArtistBody/ArtistBody'
-import { useArtistQuery } from 'features/artist/queries/useArtistQuery'
 import { UseRouteType } from 'features/navigation/navigators/RootNavigator/types'
 import { PageNotFound } from 'features/navigation/pages/PageNotFound'
 import { analytics } from 'libs/analytics/provider'
 import { eventMonitoring } from 'libs/monitoring/services'
+import { useArtistQuery } from 'queries/artist/useArtistQuery'
 import { useArtistResultsQuery } from 'queries/offer/useArtistResultsQuery'
 import { usePageTracking } from 'shared/tracking/usePageTracking'
 import { LoadingPage } from 'ui/pages/LoadingPage'
