+++ src/features/artist/pages/ArtistWebview.tsx
@@ -4,10 +4,10 @@ import { ErrorBoundary } from 'react-error-boundary'
 import { WebView } from 'react-native-webview'
 import styled from 'styled-components/native'
 
-import { useArtistQuery } from 'features/artist/queries/useArtistQuery'
 import { UseRouteType } from 'features/navigation/navigators/RootNavigator/types'
 import { PageNotFound } from 'features/navigation/pages/PageNotFound'
 import { eventMonitoring } from 'libs/monitoring/services'
+import { useArtistQuery } from 'queries/artist/useArtistQuery'
 import { useGetHeaderHeight } from 'shared/header/useGetHeaderHeight'
 import { PageHeaderWithoutPlaceholder } from 'ui/components/headers/PageHeaderWithoutPlaceholder'
 import { LoadingPage } from 'ui/pages/LoadingPage'
