+++ src/queries/artist/useArtistQuery.ts
